# Add high-precision, low-distortion, but expensive timer, as build time option.