--- a/Modules/timemodule.c
+++ b/Modules/timemodule.c
@@ -2296,3 +2296,4 @@ error:
     return -1;
 #endif
 }
+